refactor(drift_sync): extract registry.py (3/8)
Move CheckFn, register_check, _CHECKS, run_all_checks, _SEVERITY_RANK,
and _filter_by_severity into registry.py. __init__.py re-exports all
five public names (including the private _filter_by_severity used by
commands/drift.py via attribute access).

Unused imports (Callable, chain) cleaned up from __init__.py.

Refs #47 (Theme A item 4).
