+++ src/gh_manage/drift_sync/__init__.py
@@ -41,10 +41,8 @@ from __future__ import annotations
 import hashlib
 import json as _json
 import re as _re
-from collections.abc import Callable
 from datetime import datetime, timedelta
 from importlib.resources import files as _package_files
-from itertools import chain
 from pathlib import Path
 from typing import Any
 
@@ -61,53 +59,15 @@ from gh_manage.drift_sync.context import (  # noqa: F401
 )
 
 
-# ========== Check Registry ==========
+# ========== Check Registry (moved to drift_sync.registry in cli/v1.7.0) ==========
 
-
-CheckFn = Callable[["ScanContext"], tuple[Finding, ...]]
-_CHECKS: list[CheckFn] = []
-
-
-def register_check(fn: CheckFn) -> CheckFn:
-    """Decorator: register a check function in the global registry.
-
-    Intended usage:
-
-        @register_check
-        def check_labels(ctx: ScanContext) -> tuple[Finding, ...]:
-            ...
-
-    Order of registration determines order of execution in
-    run_all_checks. Phase 8 registers check_labels, check_protection,
-    check_profile_files in that order.
-    """
-    _CHECKS.append(fn)
-    return fn
-
-
-def run_all_checks(ctx: ScanContext) -> tuple[Finding, ...]:
-    """Run every registered check in order and concatenate findings.
-
-    Fail-fast: if a check raises, the exception propagates and no
-    further checks run. MVP does not have a --continue-on-error flag;
-    that is filed as a Phase 8.5+ Issue.
-    """
-    return tuple(chain.from_iterable(check(ctx) for check in _CHECKS))
-
-
-_SEVERITY_RANK = {"critical": 3, "high": 2, "medium": 1, "low": 0}
-
-
-def _filter_by_severity(
-    findings: tuple[Finding, ...], min_severity: Severity
-) -> tuple[Finding, ...]:
-    """Filter findings to those with severity >= min_severity.
-
-    Hierarchy (highest to lowest): critical > high > medium > low.
-    Input order is preserved for stable reporting.
-    """
-    threshold = _SEVERITY_RANK[min_severity]
-    return tuple(f for f in findings if _SEVERITY_RANK[f.severity] >= threshold)
+from gh_manage.drift_sync.registry import (  # noqa: E402, F401
+    _CHECKS,
+    CheckFn,
+    _filter_by_severity,
+    register_check,
+    run_all_checks,
+)
 
 
 # ========== Adapters ==========
